+++ src/main/java/com/bunq/sdk/model/core/BunqModel.java
@@ -75,10 +75,10 @@ abstract public class BunqModel {
     /**
      * De-serializes an ID object and returns its integer value.
      */
-    protected static BunqResponse<Integer> processForId(BunqResponseRaw responseRaw) {
+    protected static BunqResponse<Long> processForId(BunqResponseRaw responseRaw) {
         JsonObject responseItemObject = getResponseItemObject(responseRaw);
         JsonObject responseItemObjectUnwrapped = getWrappedContent(responseItemObject, FIELD_ID);
-        Integer responseValue = gson.fromJson(responseItemObjectUnwrapped, Id.class).getId();
+        Long responseValue = gson.fromJson(responseItemObjectUnwrapped, Id.class).getId();
 
         return new BunqResponse<>(responseValue, responseRaw.getHeaders());
     }
@@ -195,11 +195,11 @@ abstract public class BunqModel {
         return BunqContext.getApiContext();
     }
 
-    protected static Integer determineUserId() {
+    protected static Long determineUserId() {
         return BunqContext.getUserContext().getUserId();
     }
 
-    protected static Integer determineMonetaryAccountId(Integer id) {
+    protected static Long determineMonetaryAccountId(Long id) {
         if (id == null) {
             return BunqContext.getUserContext().getPrimaryMonetaryAccountBank().getId();
         } else {
